pyproject.toml
[project]
name = "python-picnic-api2"
version = "1.1.2"
description = ""
readme = "README.rst"
license = "Apache-2.0"
maintainers = [
  { name = "CodeSalat", email = "pypi@codesalat.dev"}
]
authors = [
  { name = "Mike Brink", email = "mjh.brink@icloud.com"},
  { name = "CodeSalat", email = "pypi@codesalat.dev"}
]
urls = {homepage = "https://github.com/codesalatdev/python-picnic-api", repository = "https://github.com/codesalatdev/python-picnic-api"}
requires-python = ">=3.11"
dependencies = [
  "requests>=2.24.0",
  "typing_extensions>=4.12.2"
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
ruff = "^0.9.6"
python-dotenv = "^0.15.0"

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py311"

[tool.ruff.lint]
select = [
  # pycodestyle
  "E",
  # Pyflakes
  "F",
  # pyupgrade
  "UP",
  # flake8-bugbear
  "B",
  # flake8-simplify
  "SIM",
  # isort
  "I",
  # flake8-fixme
  "FIX"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
